+++ main.tf
@@ -1,32 +1,30 @@
 ########## Start AWS VPC and Network Configurations Resources ##########
-# # ## VPC resource - Create only if vpc_id is not provided
-# locals {
-#   # Determine if a new VPC should be created
-#   create_vpc = var.vpc_id == null
-#   create_igw = var.internet_gateway_id == null
-# }
-# resource "aws_vpc" "this" {
-#   for_each = local.create_vpc ? { "vpc" = true } : {}
+// Create VPC
+resource "aws_vpc" "cato-vpc" {
+  count      = var.vpc_id == null ? 1 : 0
+  cidr_block = var.vpc_range
+  tags = {
+    Name = "${var.site_name}-VPC"
+  }
+}
 
-#   cidr_block = var.vpc_range
-#   tags = {
-#     Name = "${var.site_name}-VPC"
-#   }
-# }
-# data "aws_vpc" "existing" {
-#   count = var.vpc_id != null ? 1 : 0
-#   id    = var.vpc_id
-# }
+# Internet Gateway
+resource "aws_internet_gateway" "internet_gateway" {
+  count = var.internet_gateway_id == null ? 1 : 0
+  tags = {
+    Name = "${var.site_name}-IGW"
+  }
+  vpc_id = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
+}
 
 # Lookup data from region and VPC - Always needed for availability zones
-data "aws_availability_zones" "available" {
+data "aws_availability_zones" "available_zones" {
   state = "available"
 }
 
 # Subnets
 resource "aws_subnet" "mgmt_subnet" {
-  # vpc_id            = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id            = var.vpc_id
+  vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   cidr_block        = var.subnet_range_mgmt
   availability_zone = data.aws_availability_zones.available.names[0]
   tags = {
@@ -35,8 +33,7 @@ resource "aws_subnet" "mgmt_subnet" {
 }
 
 resource "aws_subnet" "wan_subnet" {
-  # vpc_id            = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id            = var.vpc_id
+  vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   cidr_block        = var.subnet_range_wan
   availability_zone = data.aws_availability_zones.available.names[0]
   tags = {
@@ -45,8 +42,7 @@ resource "aws_subnet" "wan_subnet" {
 }
 
 resource "aws_subnet" "lan_subnet_primary" {
-  # vpc_id            = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id            = var.vpc_id
+  vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   cidr_block        = var.subnet_range_lan_primary
   availability_zone = data.aws_availability_zones.available.names[0]
   tags = {
@@ -55,8 +51,7 @@ resource "aws_subnet" "lan_subnet_primary" {
 }
 
 resource "aws_subnet" "lan_subnet_secondary" {
-  # vpc_id            = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id            = var.vpc_id
+  vpc_id            = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   cidr_block        = var.subnet_range_lan_secondary
   availability_zone = data.aws_availability_zones.available.names[0]
   tags = {
@@ -64,20 +59,18 @@ resource "aws_subnet" "lan_subnet_secondary" {
   }
 }
 
-
 # Internal and External Security Groups
 resource "aws_security_group" "internal_sg" {
   name        = "${var.site_name}-Internal-SG"
   description = "CATO LAN Security Group - Allow all traffic Inbound"
-  # vpc_id            = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id = var.vpc_id
+  vpc_id      = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   ingress = [
     {
       description      = "Allow all traffic Inbound from Ingress CIDR Blocks"
       protocol         = -1
       from_port        = 0
       to_port          = 0
-      cidr_blocks      = var.ingress_cidr_blocks
+      cidr_blocks      = var.lan_ingress_cidr_blocks
       ipv6_cidr_blocks = []
       prefix_list_ids  = []
       security_groups  = []
@@ -102,11 +95,10 @@ resource "aws_security_group" "internal_sg" {
   }
 }
 
-resource "aws_security_group" "external_sg" {
-  name        = "${var.site_name}-External-SG"
-  description = "CATO WAN Security Group - Allow HTTPS In"
-  # vpc_id            = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id = var.vpc_id
+resource "aws_security_group" "external_sg_mgmt" {
+  name        = "${var.site_name}-External-SG-MGMT"
+  description = "CATO MGMT Security Group - Allow HTTPS In"
+  vpc_id      = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   ingress = [
     {
       description      = "Allow HTTPS In"
@@ -145,7 +137,29 @@ resource "aws_security_group" "external_sg" {
     }
   ]
   tags = {
-    name = "${var.site_name}-External-SG"
+    name = "${var.site_name}-External-SG-MGMT"
+  }
+}
+
+resource "aws_security_group" "external_sg_wan" {
+  name        = "${var.site_name}-External-SG-WAN"
+  description = "CATO WAN Security Group - Allow all out, none in"
+  vpc_id      = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
+  egress = [
+    {
+      description      = "Allow all traffic Outbound"
+      protocol         = -1
+      from_port        = 0
+      to_port          = 0
+      cidr_blocks      = ["0.0.0.0/0"]
+      ipv6_cidr_blocks = []
+      prefix_list_ids  = []
+      security_groups  = []
+      self             = false
+    }
+  ]
+  tags = {
+    name = "${var.site_name}-External-SG-WAN"
   }
 }
 
@@ -154,7 +168,7 @@ resource "aws_network_interface" "mgmteni_primary" {
   source_dest_check = "true"
   subnet_id         = aws_subnet.mgmt_subnet.id
   private_ips       = [var.mgmt_eni_primary_ip]
-  security_groups   = [aws_security_group.external_sg.id]
+  security_groups   = [aws_security_group.external_sg_mgmt.id]
   tags = {
     Name = "${var.site_name}-MGMT-INT-Primary"
   }
@@ -164,7 +178,7 @@ resource "aws_network_interface" "mgmteni_secondary" {
   source_dest_check = "true"
   subnet_id         = aws_subnet.mgmt_subnet.id
   private_ips       = [var.mgmt_eni_secondary_ip]
-  security_groups   = [aws_security_group.external_sg.id]
+  security_groups   = [aws_security_group.external_sg_mgmt.id]
   tags = {
     Name = "${var.site_name}-MGMT-INT-Secondary"
   }
@@ -174,7 +188,7 @@ resource "aws_network_interface" "waneni_primary" {
   source_dest_check = "true"
   subnet_id         = aws_subnet.wan_subnet.id
   private_ips       = [var.wan_eni_primary_ip]
-  security_groups   = [aws_security_group.external_sg.id]
+  security_groups   = [aws_security_group.external_sg_wan.id]
   tags = {
     Name = "${var.site_name}-WAN-INT-Primary"
   }
@@ -184,7 +198,7 @@ resource "aws_network_interface" "waneni_secondary" {
   source_dest_check = "true"
   subnet_id         = aws_subnet.wan_subnet.id
   private_ips       = [var.wan_eni_secondary_ip]
-  security_groups   = [aws_security_group.external_sg.id]
+  security_groups   = [aws_security_group.external_sg_wan.id]
   tags = {
     Name = "${var.site_name}-WAN-INT-Secondary"
   }
@@ -259,24 +273,21 @@ resource "aws_eip_association" "mgmteip_assoc_secondary" {
 
 # Routing Tables
 resource "aws_route_table" "wanrt" {
-  # vpc_id            = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id = var.vpc_id
+  vpc_id = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   tags = {
     Name = "${var.site_name}-WAN-RT"
   }
 }
 
 resource "aws_route_table" "mgmtrt" {
-  # vpc_id            = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id = var.vpc_id
+  vpc_id = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   tags = {
     Name = "${var.site_name}-MGMT-RT"
   }
 }
 
 resource "aws_route_table" "lanrt" {
-  # vpc_id            = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id = var.vpc_id
+  vpc_id = var.vpc_id == null ? aws_vpc.cato-vpc[0].id : var.vpc_id
   tags = {
     Name = "${var.site_name}-LAN-RT"
   }
@@ -286,15 +297,13 @@ resource "aws_route_table" "lanrt" {
 resource "aws_route" "wan_route" {
   route_table_id         = aws_route_table.wanrt.id
   destination_cidr_block = "0.0.0.0/0"
-  # gateway_id             = local.create_igw ? aws_internet_gateway.this["internet_gateway"].id : var.internet_gateway_id
-  gateway_id = var.internet_gateway_id
+  gateway_id             = var.internet_gateway_id == null ? aws_internet_gateway.internet_gateway[0].id : var.internet_gateway_id
 }
 
 resource "aws_route" "mgmt_route" {
   route_table_id         = aws_route_table.mgmtrt.id
   destination_cidr_block = "0.0.0.0/0"
-  # gateway_id             = local.create_igw ? aws_internet_gateway.this["internet_gateway"].id : var.internet_gateway_id
-  gateway_id = var.internet_gateway_id
+  gateway_id             = var.internet_gateway_id == null ? aws_internet_gateway.internet_gateway[0].id : var.internet_gateway_id
 }
 
 resource "aws_route" "lan_route" {
@@ -328,34 +337,243 @@ resource "aws_route_table_association" "lan_subnet_route_table_association_secon
 
 ########## Start Cato Site and Vsocket Deployment Resources ##########
 
-module "vsocket-aws-ha" {
-  source     = "catonetworks/vsocket-aws-ha/cato"
-  token      = var.token
-  account_id = var.account_id
-  # vpc_id                         = local.create_vpc ? aws_vpc.this["vpc"].id : var.vpc_id
-  vpc_id                         = var.vpc_id
-  key_pair                       = var.key_pair
-  region                         = var.region
-  site_name                      = var.site_name
-  site_description               = var.site_description
-  site_type                      = var.site_type
-  native_network_range_primary   = var.subnet_range_lan_primary
-  native_network_range_secondary = var.subnet_range_lan_secondary
-  mgmt_subnet_id                 = aws_subnet.mgmt_subnet.id
-  wan_subnet_id                  = aws_subnet.wan_subnet.id
-  lan_local_primary_ip           = var.lan_eni_primary_ip
-  lan_local_secondary_ip         = var.lan_eni_secondary_ip
-  lan_subnet_primary_id          = aws_subnet.lan_subnet_primary.id
-  lan_subnet_secondary_id        = aws_subnet.lan_subnet_secondary.id
-  mgmt_eni_primary_id            = aws_network_interface.mgmteni_primary.id
-  wan_eni_primary_id             = aws_network_interface.waneni_primary.id
-  lan_eni_primary_id             = aws_network_interface.laneni_primary.id
-  mgmt_eni_secondary_id          = aws_network_interface.mgmteni_secondary.id
-  wan_eni_secondary_id           = aws_network_interface.waneni_secondary.id
-  lan_eni_secondary_id           = aws_network_interface.laneni_secondary.id
-  lan_route_table_id             = aws_route_table.lanrt.id
-  site_location                  = var.site_location
-  tags                           = var.tags
+resource "cato_socket_site" "aws-site" {
+  connection_type = var.connection_type
+  description     = var.site_description
+  name            = var.site_name
+  native_range = {
+    native_network_range = var.subnet_range_lan_primary
+    local_ip             = var.lan_eni_primary_ip
+  }
+  site_location = var.site_location
+  site_type     = var.site_type
+}
+
+data "cato_accountSnapshotSite" "aws-site-primary" {
+  id = cato_socket_site.aws-site.id
+}
+
+locals {
+  primary_serial = [for s in data.cato_accountSnapshotSite.aws-site-primary.info.sockets : s.serial if s.is_primary == true]
+  sanitized_name = replace(replace(replace(replace(replace(replace(
+    var.site_name,
+    "/", ""),
+    ":", ""),
+    "#", ""),
+    "(", ""),
+    ")", ""),
+  " ", "-")
+}
+
+# AWS HA IAM role configuration
+resource "aws_iam_role" "cato_ha_role" {
+  name        = "Cato-HA-Role-${local.sanitized_name}"
+  description = "To allow vSocket HA route management"
+  assume_role_policy = jsonencode({
+    Version = "2012-10-17",
+    Statement = [{
+      Effect = "Allow"
+      Action = "sts:AssumeRole"
+      Principal = {
+        Service = "ec2.amazonaws.com"
+      }
+    }]
+  })
+}
+
+resource "aws_iam_policy" "cato_ha_policy" {
+  name = "Cato-HA-Role-Policy-${local.sanitized_name}"
+  policy = jsonencode({
+    Version = "2012-10-17",
+    Statement = [
+      {
+        Effect = "Allow"
+        Action = [
+          "ec2:CreateRoute",
+          "ec2:DescribeRouteTables",
+          "ec2:ReplaceRoute"
+        ]
+        Resource = "*"
+      },
+    ]
+  })
+}
+
+resource "aws_iam_role_policy_attachment" "cato_ha_attach" {
+  role       = aws_iam_role.cato_ha_role.name
+  policy_arn = aws_iam_policy.cato_ha_policy.arn
+}
+
+resource "aws_iam_instance_profile" "cato_ha_instance_profile" {
+  name = "Cato-HA-Role-${local.sanitized_name}"
+  role = aws_iam_role.cato_ha_role.name
+}
+
+## Lookup data from region and VPC
+data "aws_ami" "vsocket" {
+  most_recent = true
+  name_regex  = "VSOCKET_AWS"
+  owners      = ["aws-marketplace"]
+}
+
+data "aws_availability_zones" "available" {
+  state = "available"
+}
+
+# Create Primary vSocket Virtual Machine
+resource "aws_instance" "primary_vsocket" {
+  tenancy              = "default"
+  ami                  = data.aws_ami.vsocket.id
+  key_name             = var.key_pair
+  instance_type        = var.instance_type
+  user_data            = base64encode(local.primary_serial[0])
+  iam_instance_profile = aws_iam_instance_profile.cato_ha_instance_profile.name
+  # Network Interfaces
+  # MGMTENI
+  network_interface {
+    device_index         = 0
+    network_interface_id = aws_network_interface.mgmteni_primary.id
+  }
+  # WANENI
+  network_interface {
+    device_index         = 1
+    network_interface_id = aws_network_interface.waneni_primary.id
+  }
+  # LANENI
+  network_interface {
+    device_index         = 2
+    network_interface_id = aws_network_interface.laneni_primary.id
+  }
+  ebs_block_device {
+    device_name = "/dev/sda1"
+    volume_size = var.ebs_disk_size
+    volume_type = var.ebs_disk_type
+  }
+  tags = merge(var.tags, {
+    Name = "${var.site_name}-vSocket-Primary"
+  })
+}
+
+# To allow socket to upgrade so secondary socket can be added
+resource "null_resource" "sleep_300_seconds" {
+  provisioner "local-exec" {
+    command = "sleep 300"
+  }
+  depends_on = [aws_instance.primary_vsocket]
+}
+
+#################################################################################
+# Add secondary socket to site via API until socket_site resource is updated to natively support
+resource "null_resource" "configure_secondary_aws_vsocket" {
+  depends_on = [null_resource.sleep_300_seconds]
+
+  provisioner "local-exec" {
+    command = <<EOF
+      # Execute the GraphQL mutation to get add the secondary vSocket
+      response=$(curl -k -X POST \
+        -H "Accept: application/json" \
+        -H "Content-Type: application/json" \
+        -H "x-API-Key: ${var.token}" \
+        "${var.baseurl}" \
+        --data '{
+          "query": "mutation siteAddSecondaryAwsVSocket($accountId: ID!, $addSecondaryAwsVSocketInput: AddSecondaryAwsVSocketInput!) { site(accountId: $accountId) { addSecondaryAwsVSocket(input: $addSecondaryAwsVSocketInput) { id } } }",
+          "variables": {
+            "accountId": "${var.account_id}",
+            "addSecondaryAwsVSocketInput": {
+              "eniIpAddress": "${var.lan_eni_secondary_ip}",
+              "eniIpSubnet": "${var.subnet_range_lan_secondary}",
+               "routeTableId": "${aws_route_table.lanrt.id}",
+              "site": {
+                "by": "ID",
+                "input": "${cato_socket_site.aws-site.id}"
+              }
+            }
+          },
+          "operationName": "siteAddSecondaryAwsVSocket"
+        }' )
+    EOF
+  }
+
+  triggers = {
+    account_id = var.account_id
+    site_id    = cato_socket_site.aws-site.id
+  }
+}
+
+# Sleep to allow Secondary vSocket serial retrieval
+resource "null_resource" "sleep_30_seconds" {
+  provisioner "local-exec" {
+    command = "sleep 30"
+  }
+  depends_on = [null_resource.configure_secondary_aws_vsocket]
+}
+
+# Retrieve Secondary vSocket Virtual Machine serial
+data "cato_accountSnapshotSite" "aws-site-secondary" {
+  depends_on = [null_resource.sleep_30_seconds]
+  id         = cato_socket_site.aws-site.id
+}
+
+locals {
+  secondary_serial = [for s in data.cato_accountSnapshotSite.aws-site-secondary.info.sockets : s.serial if s.is_primary == false]
+  depends_on       = [data.cato_accountSnapshotSite.aws-site-secondary]
+}
+
+## vSocket Instance
+resource "aws_instance" "secondary_vsocket" {
+  tenancy              = "default"
+  ami                  = data.aws_ami.vsocket.id
+  key_name             = var.key_pair
+  instance_type        = var.instance_type
+  user_data            = base64encode(local.secondary_serial[0])
+  iam_instance_profile = aws_iam_instance_profile.cato_ha_instance_profile.name
+  # Network Interfaces
+  # MGMTENI
+  network_interface {
+    device_index         = 0
+    network_interface_id = aws_network_interface.mgmteni_secondary.id
+  }
+  # WANENI
+  network_interface {
+    device_index         = 1
+    network_interface_id = aws_network_interface.waneni_secondary.id
+  }
+  # LANENI
+  network_interface {
+    device_index         = 2
+    network_interface_id = aws_network_interface.laneni_secondary.id
+  }
+  ebs_block_device {
+    device_name = "/dev/sda1"
+    volume_size = var.ebs_disk_size
+    volume_type = var.ebs_disk_type
+  }
+  tags = merge(var.tags, {
+    Name = "${var.site_name}-vSocket-Secondary"
+  })
+  depends_on = [null_resource.sleep_30_seconds]
+}
+
+# To allow sockets to configure HA
+resource "null_resource" "sleep_300_seconds-HA" {
+  provisioner "local-exec" {
+    command = "sleep 300"
+  }
+  depends_on = [aws_instance.secondary_vsocket]
+}
+
+data "cato_accountSnapshotSite" "aws-site-2" {
+  id         = cato_socket_site.aws-site.id
+  depends_on = [null_resource.sleep_300_seconds-HA]
+}
+
+resource "cato_license" "license" {
+  depends_on = [aws_instance.secondary_vsocket]
+  count      = var.license_id == null ? 0 : 1
+  site_id    = cato_socket_site.aws-site.id
+  license_id = var.license_id
+  bw         = var.license_bw == null ? null : var.license_bw
 }
 
 ########## End Cato Site and Vsocket Deployment Resources ##########
+